remove unused sections from app, extract out date string creator
