+++ src/js/libs/date.js
@@ -1,0 +1,9 @@
+const generateDateString = () => {
+  const date = new Date();
+  const year = date.getFullYear();
+  const month = (date.getMonth() < 9) ? "0" + (date.getMonth() + 1) : (date.getMonth() + 1);
+  const day = (date.getDate() < 10) ? "0" + date.getDate() : date.getDate();
+  return year + '-' + month + '-' + day;
+};
+
+export { generateDateString };
